fix: 修复多项代码缺陷并优化项目质量
本次提交修复了大量代码问题，包括：
1.  修复多处整数/浮点数精度不匹配问题，统一使用浮点运算
2.  优化析构函数调用安全性，添加clang-tidy注释说明
3.  重构解码栈内存管理，移除Annotation内存池，改用值类型存储
4.  修复条件判断逻辑拆分，提升代码可读性
5.  替换atoi为strtol，修复类型转换安全问题
6.  修复指针类型不匹配问题，统一const正确性
7.  修复递归锁问题，将普通mutex替换为recursive_mutex避免死锁
8.  添加clang-tidy检查脚本与配置文件，优化代码规范检查流程
9.  修复文件操作内存泄漏与错误返回值问题
10. 移除_delete_flag标记，改用shared_ptr管理解码栈生命周期

+++ PXView/pv/view/renderer/viewport_painter.cpp
@@ -359,9 +359,9 @@ void ViewportPainter::paintProgress(QPainter &p, QColor fore, QColor back) {
   if (!_viewport->transfer_started()) {
     const int width = _viewport->view().get_view_width();
     const QPoint cenLeftPos =
-        QPoint(width / 2 - 0.05 * width, _viewport->height() / 2);
+        QPoint(static_cast<int>(width / 2.0 - 0.05 * width), _viewport->height() / 2);
     const QPoint cenRightPos =
-        QPoint(width / 2 + 0.05 * width, _viewport->height() / 2);
+        QPoint(static_cast<int>(width / 2.0 + 0.05 * width), _viewport->height() / 2);
     const int trigger_radius = min(0.02 * width, 0.02 * _viewport->height());
 
     QColor foreBack = fore;
